+++ fbw-a380x/src/systems/instruments/src/MsfsAvionicsCommon/EcamMessages/index.tsx
@@ -3,18 +3,26 @@
 
 export const WD_NUM_LINES = 17;
 
-import { EcamAbnormalSensedAta212223 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata21-22-23';
+import {
+  EcamAbnormalSensedAta212223,
+  EcamDeferredProcAta212223,
+} from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata21-22-23';
 import { EcamAbnormalSensedAta24 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata24';
 import { EcamAbnormalSensedAta26 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata26';
 import { EcamAbnormalSensedAta27 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata27';
 import { EcamAbnormalSensedAta28 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata28';
 import { EcamAbnormalSensedAta2930 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata29-30';
-import { EcamAbnormalSensedAta313233 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata31-32-33';
+import {
+  EcamAbnormalSensedAta313233,
+  EcamDeferredProcAta313233,
+} from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata31-32-33';
 import { EcamAbnormalSensedAta34 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata34';
 import { EcamAbnormalSensedAta353642 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata35-36-42';
 import { EcamAbnormalSensedAta46495256 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata46-49-52-56';
 import { EcamAbnormalSensedAta70 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata70';
 import { EcamAbnormalSensedAta80Rest } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata80-rest';
+import { EcamAbnormalSecondaryFailures } from 'instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/secondary-failures';
+import { AbnormalNonSensedCategory } from 'instruments/src/MsfsAvionicsCommon/providers/FwsEwdPublisher';
 
 // Convention for IDs:
 // First two digits: ATA chapter. 00 for T.O and LDG memos
@@ -58,23 +66,12 @@ export const EcamMemos: { [n: string]: string } = {
   '322000001': '\x1b<4mN/W STEER DISC',
   '322000002': '\x1b<3mN/W STEER DISC',
   '000005001': '\x1b<3mREFUELG',
-  '000005501': '\x1b<3mGND SPLRs ARMED',
-  '000056101': '\x1b<3mCOMPANY ALERT',
-  '000056102': '\x1b<3m\x1b)mCOMPANY ALERT',
-  '000006001': '\x1b<3mSPEED BRK',
-  '000006002': '\x1b<4mSPEED BRK',
-  '000010501': '\x1b<3mOUTR TK FUEL XFRD',
-  '000011001': '\x1b<3mFOB BELOW 3 T',
-  '000011002': '\x1b<3mFOB BELOW 6600 LBS',
-  '000013501': '\x1b<3mACARS STBY',
-  '000066001': '\x1b<3mGSM DISC < 4MN',
   '290000001': '\x1b<3mG ELEC PMP A CTL',
   '290000002': '\x1b<3mG ELEC PMP B CTL',
   '290000003': '\x1b<3mY ELEC PMP A CTL',
   '290000004': '\x1b<3mY ELEC PMP B CTL',
   '000017001': '\x1b<3mAPU AVAIL',
   '000018001': '\x1b<3mAPU BLEED',
-  '000019001': '\x1b<3mLDG LT',
   '240000001': '\x1b<3mCOMMERCIAL PART SHED',
   '241000001': '\x1b<4mELEC EXT PWR',
   '241000002': '\x1b<3mELEC EXT PWR',
@@ -83,9 +80,6 @@ export const EcamMemos: { [n: string]: string } = {
   '243000001': '\x1b<3mREMOTE C/B CTL ON',
   '000023001': '\x1b<3mMAN LDG ELEV',
   '000029001': '\x1b<3mSWITCHG PNL',
-  '000035001': '\x1b<2mLAND ASAP',
-  '000036001': '\x1b<4mLAND ASAP',
-  '000056001': '\x1b<3mHI ALT SET',
   '220000001': '\x1b<2mAP OFF',
   '220000002': '\x1b<4mA/THR OFF',
   '221000001': '\x1b<3mFMS SWTG',
@@ -179,19 +173,7 @@ export const EcamMemos: { [n: string]: string } = {
 };
 
 /** Only these IDs will be shown in the PFD MEMO section */
-export const pfdMemoDisplay: string[] = [
-  '322000001',
-  '000006002',
-  '000032501',
-  '000035001',
-  '000036001',
-  '000054502',
-  '220000001',
-  '220000002',
-  '300000001',
-  '300000002',
-  '320000001',
-];
+export const pfdMemoDisplay: string[] = ['000006002', '220000001', '220000002', '300000001', '300000002', '320000001'];
 
 /** All possible INFOs (e.g. CAT 3 SINGLE ONLY), with special formatting characters. */
 export const EcamInfos: { [n: string]: string } = {
@@ -210,6 +192,7 @@ export const EcamInfos: { [n: string]: string } = {
   220200010: '\x1b<3mCAT 1 ONLY',
   230200001: '\x1b<3mSATCOM DATALINK AVAIL',
   260200001: '\x1b<3mBEFORE CARGO OPENING : PAX DISEMBARK',
+  270200001: '\x1bON DRY RWY ONLY : LDG DIST AFFECTED < 15%',
   320200001: '\x1b<3mALTN BRK WITH A-SKID',
   320200002: '\x1b<3mBRK PRESS AUTO LIMITED ON ALL L/Gs',
   320200003: '\x1b<3mDELAY BRAKING UNTIL NLG TOUCHDOWN',
@@ -465,29 +448,48 @@ export const EcamInopSys: { [n: string]: string } = {
   260300060: '\x1b<4mU3R RCC SMK DET',
   260300061: '\x1b<4mU1L SHOWER SMK DET',
   260300062: '\x1b<4mU1R SHOWER SMK DET',
-  290300001: '\x1b<4m G ELEC PMP A',
-  290300002: '\x1b<4m G ELEC PMP B',
-  290300003: '\x1b<4m Y ELEC PMP A',
-  290300004: '\x1b<4m Y ELEC PMP B',
-  290300005: '\x1b<4m G ENG 1 PMP A',
-  290300006: '\x1b<4m G ENG 1 PMP B',
-  290300007: '\x1b<4m G ENG 2 PMP A',
-  290300008: '\x1b<4m G ENG 2 PMP B',
-  290300009: '\x1b<4m Y ENG 3 PMP A',
-  290300010: '\x1b<4m Y ENG 3 PMP B',
-  290300011: '\x1b<4m Y ENG 4 PMP A',
-  290300012: '\x1b<4m Y ENG 4 PMP B',
-  290300013: '\x1b<4m G SYS CHAN A OVHT DET',
-  290300014: '\x1b<4m G SYS CHAN B OVHT DET',
-  290300015: '\x1b<4m Y SYS CHAN A OVHT DET',
-  290300016: '\x1b<4m Y SYS CHAN B OVHT DET',
-  290300017: '\x1b<4m G HSMU',
-  290300018: '\x1b<4m Y HSMU',
-  290300019: '\x1b<4m G SYS OVHT DET',
-  290300020: '\x1b<4m Y SYS OVHT DET',
-  290300021: '\x1b<4m G HYD SYS',
-  290300022: '\x1b<4m Y HYD SYS',
+  270300001: '\x1b<4mSEC 1',
+  270300002: '\x1b<4mSEC 2',
+  270300003: '\x1b<4mSEC 3',
+  270300004: '\x1b<4mRUDDER TRIM',
+  270300005: '\x1b<4mRUDDER TRIM 1',
+  270300006: '\x1b<4mRUDDER TRIM 2',
+  290100001: '\x1b<4mPART SPLRs',
+  290100003: '\x1b<4mFLAPS SYS 1',
+  290100004: '\x1b<4mFLAPS SYS 2',
+  290100005: '\x1b<4mSLATS SYS 1',
+  290100006: '\x1b<4mSLATS SYS 2',
+  290100007: '\x1b<4mSTABILIZER',
+  290100008: '\x1b<4mF/CTL PROT',
+  290100009: '\x1b<4mL OUTR AILERON',
+  290100010: '\x1b<4mR OUTR AILERON',
+  290100011: '\x1b<4mMOST SPLRs',
+  290100012: '\x1b<4mFLAPS',
+  290300001: '\x1b<4mG ELEC PMP A',
+  290300002: '\x1b<4mG ELEC PMP B',
+  290300003: '\x1b<4mY ELEC PMP A',
+  290300004: '\x1b<4mY ELEC PMP B',
+  290300005: '\x1b<4mG ENG 1 PMP A',
+  290300006: '\x1b<4mG ENG 1 PMP B',
+  290300007: '\x1b<4mG ENG 2 PMP A',
+  290300008: '\x1b<4mG ENG 2 PMP B',
+  290300009: '\x1b<4mY ENG 3 PMP A',
+  290300010: '\x1b<4mY ENG 3 PMP B',
+  290300011: '\x1b<4mY ENG 4 PMP A',
+  290300012: '\x1b<4mY ENG 4 PMP B',
+  290300013: '\x1b<4mG SYS CHAN A OVHT DET',
+  290300014: '\x1b<4mG SYS CHAN B OVHT DET',
+  290300015: '\x1b<4mY SYS CHAN A OVHT DET',
+  290300016: '\x1b<4mY SYS CHAN B OVHT DET',
+  290300017: '\x1b<4mG HSMU',
+  290300018: '\x1b<4mY HSMU',
+  290300019: '\x1b<4mG SYS OVHT DET',
+  290300020: '\x1b<4mY SYS OVHT DET',
+  290300021: '\x1b<4mG HYD SYS',
+  290300022: '\x1b<4mY HYD SYS',
   310300001: '\x1b<4mAUTO CALLOUT',
+  310300002: '\x1b<4mFWS 1',
+  310300003: '\x1b<4mFWS 2',
   320300001: '\x1b<4mA-SKID',
   320300002: '\x1b<4mAUTO BRK',
   320300003: '\x1b<4mPART A-SKID',
@@ -510,6 +512,8 @@ export const EcamInopSys: { [n: string]: string } = {
   320300020: '\x1b<4mNORM N/W STEER',
   320300021: '\x1b<4mPEDAL STEER CTL',
   320300022: '\x1b<4mROW/ROP',
+  320300023: '\x1b<4mPART L/G RETRACTION',
+  320300024: '\x1b<4mNORM B/W STEER',
   340300001: '\x1b<4mGPWS 1',
   340300002: '\x1b<4mGPWS 2',
   340300003: '\x1b<4mGPWS 1+2',
@@ -559,19 +563,27 @@ export enum ChecklistLineStyle {
   Cyan = 'Cyan',
   Green = 'Green',
   Amber = 'Amber',
+  White = 'White',
   Headline = 'Headline',
   SubHeadline = 'SubHeadline',
   SeparationLine = 'SeparationLine',
   ChecklistItem = 'ChecklistItem',
+  ChecklistItemInactive = 'ChecklistItemInactive',
   CompletedChecklist = 'CompletedChecklist',
+  CompletedDeferredProcedure = 'CompletedDeferredProcedure',
+  DeferredProcedure = 'DeferredProcedure',
+  OmissionDots = 'OmissionDots',
+  LandAsap = 'LandAsap',
+  LandAnsa = 'LandAnsa',
+  ChecklistCondition = 'ChecklistCondition',
 }
 
 interface AbstractChecklistItem {
-  /** The name of the item, displayed at the beginning of the line. Does not accept special formatting tokens. No leading dot. */
+  /** The name of the item, displayed at the beginning of the line. Does not accept special formatting tokens. No leading dot. For conditions, don't include the leading "IF" */
   name: string;
   /** Sensed or not sensed item. Sensed items are automatically checked. Non-sensed items will have a checkbox drawn in front of them on the EWD */
   sensed: boolean;
-  /** On which level of indentation to print the item. 0 equals the first level. Optional, not set means first level. */
+  /** On which level of indentation to print the item. 0 equals the first level. Optional, not set means first level. Important for items subordinated to conditions. */
   level?: number;
   /** Manually define style. standard (cyan when not completed, white/green when completed), or always cyan/green/amber. Standard, if not set. */
   style?: ChecklistLineStyle;
@@ -585,7 +597,26 @@ export interface ChecklistAction extends AbstractChecklistItem {
   colonIfCompleted?: boolean;
 }
 
-interface ChecklistCondition extends AbstractChecklistItem {}
+export interface ChecklistCondition extends AbstractChecklistItem {
+  /** If this line is a condition. Can be sensed or not sensed (i.e. manually activated). */
+  condition: true;
+}
+
+export interface ChecklistSpecialItem extends AbstractChecklistItem {}
+
+export function isChecklistAction(c: AbstractChecklistItem): c is ChecklistAction {
+  return (c as ChecklistAction)?.labelNotCompleted !== undefined;
+}
+
+export function isChecklistCondition(c: AbstractChecklistItem): c is ChecklistCondition {
+  return (c as ChecklistCondition)?.condition !== undefined;
+}
+
+export function isAbnormalSensedProcedure(
+  c: AbnormalProcedure | DeferredProcedure | NormalProcedure,
+): c is AbnormalProcedure {
+  return (c as AbnormalProcedure)?.recommendation !== undefined;
+}
 
 export interface AbnormalProcedure {
   /** Title of the fault, e.g. "_HYD_ G SYS PRESS LO". \n produces second line. Accepts special formatting tokens  */
@@ -593,7 +624,7 @@ export interface AbnormalProcedure {
   /** sensed or not sensed abnormal procedure */
   sensed: boolean;
   /** An array of possible checklist items. */
-  items: (ChecklistAction | ChecklistCondition)[];
+  items: (ChecklistAction | ChecklistCondition | ChecklistSpecialItem)[];
   /** LAND ASAP or LAND ANSA displayed below title? Optional, don't fill if no recommendation */
   recommendation?: 'LAND ASAP' | 'LAND ANSA';
 }
@@ -603,12 +634,31 @@ export interface NormalProcedure {
   title: string;
   /** An array of possible checklist items.. */
   items: ChecklistAction[];
-  /** Checklist is deferred, i.e. only activated by request */
-  deferred?: boolean;
+  /** Checklist is only activated by request, deactivated per default */
+  onlyActivatedByRequest?: boolean;
 }
 
-export function isChecklistAction(element: ChecklistAction | ChecklistCondition): element is ChecklistAction {
-  return 'labelNotCompleted' in element;
+export interface AbnormalNonSensedProcedure {
+  /** Title of the checklist, e.g. "BEFORE START".  */
+  title: string;
+  /** An array of possible checklist items.. */
+  items: ChecklistAction[];
+}
+
+export enum DeferredProcedureType {
+  ALL_PHASES,
+  AT_TOP_OF_DESCENT,
+  FOR_APPROACH,
+  FOR_LANDING,
+}
+export interface DeferredProcedure {
+  /** Which abnormal procedures triggers this deferred procedure */
+  fromAbnormalProcs: string[];
+  /** (optional, only used from batch 7) Title of the procedure, Accepts special formatting tokens  */
+  title?: string;
+  /** An array of possible checklist items. */
+  items: (ChecklistAction | ChecklistCondition | ChecklistSpecialItem)[];
+  type: DeferredProcedureType;
 }
 
 /** All abnormal sensed procedures (alerts, via ECL) should be here. */
@@ -625,10 +675,26 @@ export const EcamAbnormalSensedProcedures: { [n: string]: AbnormalProcedure } = 
   ...EcamAbnormalSensedAta46495256,
   ...EcamAbnormalSensedAta70,
   ...EcamAbnormalSensedAta80Rest,
+  ...EcamAbnormalSecondaryFailures,
 };
 
-/** All abnormal non-sensed procedures (via ECL) should be here. Don't start for now, format needs to be defined. */
-export const EcamAbnormalNonSensedProcedures: { [n: string]: AbnormalProcedure } = {};
+// Abnormal non-sensed are also contained in EcamAbnormalSensedProcedures
+export const EcamAbnormalProcedures: { [n: string]: AbnormalProcedure } = EcamAbnormalSensedProcedures;
+
+export const EcamAbNormalSensedSubMenuVector: AbnormalNonSensedCategory[] = [
+  'ENG',
+  'F/CTL',
+  'L/G',
+  'NAV',
+  'FUEL',
+  'MISCELLANEOUS',
+];
+
+/** All abnormal sensed procedures (alerts, via ECL) should be here. */
+export const EcamDeferredProcedures: { [n: string]: DeferredProcedure } = {
+  ...EcamDeferredProcAta212223,
+  ...EcamDeferredProcAta313233,
+};
 
 /** Used for one common representation of data defining the visual appearance of ECAM lines on the WD (for the ECL part) */
 export interface WdLineData {
@@ -641,6 +707,7 @@ export interface WdLineData {
   lastLine: boolean;
   specialLine?: WdSpecialLine;
   abnormalProcedure?: boolean;
+  originalItemIndex?: number;
 }
 
 export enum WdSpecialLine {